Move composer install logic to listener
