+++ src/Listener/PrepareSolutionListener.php
@@ -1,0 +1,62 @@
+<?php
+
+namespace PhpSchool\PhpWorkshop\Listener;
+
+use PhpSchool\PhpWorkshop\Exercise\ExerciseInterface;
+use RuntimeException;
+use Symfony\Component\Process\Process;
+
+/**
+ * Class PrepareSolutionListener
+ * @author Aydin Hassan <aydin@hotmail.co.uk>
+ */
+class PrepareSolutionListener
+{
+    /**
+     * Locations for composer executable
+     *
+     * @var array
+     */
+    private $composerLocations = [
+        'composer',
+        'composer.phar',
+        '/usr/local/bin/composer',
+        __DIR__ . '/../vendor/bin/composer',
+    ];
+
+    /**
+     * @param ExerciseInterface $exercise
+     * @param $fileName
+     */
+    public function __invoke(ExerciseInterface $exercise, $fileName)
+    {
+        $solution = $exercise->getSolution();
+
+        if ($solution->hasComposerFile()) {
+            //prepare composer deps
+            //only install if composer.lock file not available
+
+            if (!file_exists(sprintf('%s/vendor', $solution->getBaseDirectory()))) {
+                $process = new Process(
+                    sprintf('%s install --no-interaction', $this->locateComposer()),
+                    $solution->getBaseDirectory()
+                );
+                $process->run();
+            }
+        }
+    }
+
+    /**
+     * @return string
+     */
+    private function locateComposer()
+    {
+        foreach ($this->composerLocations as $location) {
+            if (file_exists($location) && is_executable($location)) {
+                return $location;
+            }
+        }
+
+        throw new RuntimeException('Composer could not be located on the system');
+    }
+}
